refactor: rename HandlerAdapter, adapter 패키지 생성 후 재배치
