+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handleradapter/HandlerAdapter.java
@@ -1,4 +1,6 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.handleradapter;
+
+import nextstep.mvc.tobe.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
